add back button to stickyHeader
resolves #192
